fixed broken tests
Two tests in src/tests/srp_phat_test.cpp used the "abs"-function on a
float, which resulted in the float getting rounded down to zero. This in
turn made the two tests always succeed as long as the float value was
smaller than one.
One of the tests now actually fails after the fix.
This is an excellent example why C-headers shouldn't be used in C++
unless it's absolutely necessary.

The second part of this commit addresses the usage of ASSERT_TRUE for
comparisons. This is a bad idea, because gtest won't tell you the exact
values of the failed comparison. Instead, the proper
comparison-assertions should be used (e.g. ASSERT_LT)