use move semantics instead of std::ifstream pointers
Using move semantics instead of using new/delete removes the need to
manage heap memory yourself and reduces the room for errors. It also
removes the need for writing destructors that only contain a delete.

Of course, this change makes the class non-copyable, which is why I
explicitly removed the copy-constructors whenever I stumbled across one
and also had to change a test-file.
However, these classes never were copyable in the first place (guess
what happens when you let a default copy-constructor copy a
std::ifstream*), the only difference is that they're now explicit about
it.
Of course, you can still construct objects of these classes directly or
via move semantics (and new/delete, but who wants that).

Another small change in this commit is the usage of the boolean operator
of std::ifstream to check whether a file was correctly opened.